Common: Move generic upload texture code from d3d12
