+++ rpcs3/Emu/RSX/GL/GLGSRender.h
@@ -18,7 +18,6 @@
 extern GLenum g_last_gl_error;
 void printGlError(GLenum err, const char* situation);
 void printGlError(GLenum err, const std::string& situation);
-u32 LinearToSwizzleAddress(u32 x, u32 y, u32 z, u32 log2_width, u32 log2_height, u32 log2_depth);
 
 
 class GLTexture
